% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_interface.R
\name{ft3_serve_interface}
\alias{ft3_serve_interface}
\title{Title}
\usage{
ft3_serve_interface(
  api_location = NULL,
  assignments_pkg = flexTeaching3.api::ft3_options("assignments_pkg"),
  http_port = 8081,
  log_options = list(level = "off")
)
}
\arguments{
\item{assignments_pkg}{}

\item{http_port}{}

\item{log_options}{}

\item{api}{location}
}
\description{
Title
}
