% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_interface.R
\name{ft3_serve_interface}
\alias{ft3_serve_interface}
\title{Title}
\usage{
ft3_serve_interface(
  api_location = NULL,
  assignments_pkg,
  http_port = 8081,
  log_options = list(level = "off"),
  ...
)
}
\arguments{
\item{api_location}{The URL for the api (without the '/ft3/api/v1/' appended)}

\item{assignments_pkg}{Package containing the assignments}

\item{http_port}{Port on which to start the server}

\item{log_options}{List of options for the log, to be passed to RestRserve::Logger$new()}

\item{...}{Further arguments to be passed to the RestRserve backend's start() method}
}
\value{
The result from the RestRserve backend's start() method
}
\description{
Title
}
